Moved some tests
